size_t counters.

+++ brainfuck/interpreter.hpp
@@ -67,15 +67,15 @@ struct _interpret_command<chr<','>, st> {
 
 CREATE_ALIAS(interpret_command);
 
-template <typename program, typename loop = list<>, int counter = 0>
+template <typename program, typename loop = list<>, size_t counter = 0>
 struct find_loop :
         public find_loop<pop<program>, append<peek<program>, loop>, counter> {};
 
-template <typename... cmds, typename loop, int counter>
+template <typename... cmds, typename loop, size_t counter>
 struct find_loop<list<chr<'['>, cmds...>, loop, counter> :
         public find_loop<list<cmds...>, append<chr<'['>, loop>, counter + 1> {};
 
-template <typename... cmds, typename loop, int counter>
+template <typename... cmds, typename loop, size_t counter>
 struct find_loop<list<chr<']'>, cmds...>, loop, counter> :
         public find_loop<list<cmds...>, append<chr<']'>, loop>, counter - 1> {};
